refactoring Location feature
