+++ src/features/Location/index.js
@@ -1,0 +1,3 @@
+import LocationDetail from "./LocationDetails";
+
+export default LocationDetail;
