+++ src/features/Map/Map.js
@@ -3,7 +3,7 @@ import 'leaflet/dist/leaflet.css';
 import 'leaflet-defaulticon-compatibility/dist/leaflet-defaulticon-compatibility.webpack.css';
 import 'leaflet-defaulticon-compatibility';
 import { MapContainer, TileLayer, Marker } from 'react-leaflet'
-import { selectFilteredLocations } from '../searchBar/locationsSlice';
+import { selectFilteredLocations } from '../Location/locationsSlice';
 import { useSelector } from 'react-redux';
 import Leaflet from 'leaflet';
 
@@ -12,7 +12,7 @@ const customMapStyle = {
   height: '100vh'
 };
 
-function Map({children}) {
+export default function Map({children}) {
   const locations = useSelector(selectFilteredLocations);
 
   return ( 
@@ -21,5 +21,3 @@ function Map({children}) {
       </MapContainer>
     );
 }
-
-export default Map;
