+++ src/features/Map/index.js
@@ -1,4 +1,4 @@
-import dynamic from "next/dynamic";
+import dynamic from 'next/dynamic';
 
 const Map = dynamic(() => import('./Map'), {
     ssr: false
