+++ src/features/SearchBar/SearchBar.js
@@ -1,29 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import Image from 'next/image';
-import { filterLocations, selectFilteredLocations, fetchLocations, selectSelectedLocation, setLocation } from './locationsSlice';
+import { filterLocations, selectFilteredLocations, fetchLocations, selectSelectedLocation, setLocation } from '../Location/locationsSlice';
 import { useSelector, useDispatch } from 'react-redux';
-
-export const ItemList = ({item, style="", specialIcon=false, handlerClick=()=>{}}) => {
-
-  return (
-      <div className={"flex p-2 " + style} onClick={handlerClick}>
-        <Image src={specialIcon? "icon-pin-gold.svg":"icon-pin.svg"} 
-          alt="Pin Icon" width={17} height={17} className="ml-2 mr-3" />
-
-        <div >
-            <p className='text-sm font-bold'>{item.name}</p>
-            <p className='text-sm text-gray-400'>{`${item.location.lat}, ${item.location.lon}`}</p>
-
-        </div>
-      </div>
-  );
-}
-
-const ResultList = ({list}) => {
-  const dispatch = useDispatch();
-  const selectedLocation = useSelector(selectSelectedLocation);
+import LocationOverview from '@/features/Location/LocationOverview';
 
 
+const ResultList = ({list, selectedItem, handlerClick}) => {  
   return (
     <>
       {list.length===0?
@@ -35,14 +17,14 @@ const ResultList = ({list}) => {
           <p className='text-sm text-gray-50 ml-2'>Found {list.length} Results:</p>
         </div>
       }
-      <div classname='mb-2'>
+      <div className='mb-2'>
         {list.map( item => (
-          <ItemList 
+          <LocationOverview 
             key={item.id} 
             item={item} 
             style="hover:bg-blue-100"
-            specialIcon={selectedLocation != null && item.id === selectedLocation.id} 
-            handlerClick={() => dispatch(setLocation(item))}
+            specialIcon={selectedItem !== null && item.id === selectedItem.id} 
+            handlerClick={() => handlerClick(item)}
           />)
         )}
       </div>
@@ -56,19 +38,13 @@ export default function SearchBar() {
     setInput(newValue);
     dispatch(filterLocations(newValue));
   };
-
+  
   const [input, setInput] = useState('');
-
+  
   const dispatch = useDispatch();
+  const selectedLocation = useSelector(selectSelectedLocation);
   const locations = useSelector(selectFilteredLocations);
-  const locationsStatus = useSelector(state => state.locations.status);
-  const error = useSelector(state => state.locations.error);
 
-  useEffect(() => {
-    if (locationsStatus === 'idle') {
-      dispatch(fetchLocations());
-    }
-  }, [locationsStatus, dispatch]);
 
   return (
     <div className="fixed left-4 top-4 w-96 text-black">
@@ -85,7 +61,7 @@ export default function SearchBar() {
       </div>
 
       <div className='mt-4 w-full rounded bg-gray-50 shadow-md' >
-        {input!='' && <ResultList list={locations}/>}
+        {input!='' && <ResultList list={locations} selectedItem={selectedLocation} handlerClick={(item) => dispatch(setLocation(item))}/>}
       </div>
     </div>
   );
