+++ src/redux/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import locationsReducer from '@/features/searchBar/locationsSlice';
+import locationsReducer from '@/features/Location/locationsSlice';
 
 export default configureStore({
   reducer: {
